requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
cssselect2==0.7.0
DateTime==5.1
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
django_debug_toolbar==3.8.1
easy-thumbnails==2.8.5
filelock==3.9.0
fonttools==4.39.3
html5lib==1.1
identify==2.5.18
kombu==5.2.4
nodeenv==1.7.0
packaging==23.1
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.1.1
prompt-toolkit==3.0.37
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
pytesseract==0.3.10
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
vine==5.0.0
virtualenv==20.20.0
wcwidth==0.2.6
weasyprint==58.1
webencodings==0.5.1
zope.interface==6.0
zopfli==0.2.2
